pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Multi-stage document intelligence pipeline for heterogeneous PDF extraction with provenance"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    "docling>=2.0",
    "langdetect>=1.0",
    "pyyaml>=6.0",
    "pytest>=8.0",
]
[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
triage = "src.cli.triage:main"
extract = "src.cli.extract:main"
